添加ES6初期版本
