+++ package.json
@@ -1,29 +1,46 @@
 {
   "name": "xjs-framework",
-  "version": "1.0.0",
-  "private": true,
+  "version": "2.0.0",
+  "description": "a front-end framework with ES6",
+  "main": "gulpfile.js",
+  "directories": {
+    "doc": "docs"
+  },
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/a2604882741z/xjs-framework.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/a2604882741z/xjs-framework/issues"
+  },
+  "homepage": "https://github.com/a2604882741z/xjs-framework#readme",
   "devDependencies": {
-    "gulp": "^3.8.11",
-    "gulp-cache": "^0.4.5",
-    "gulp-clean": "^0.3.1",
-    "gulp-concat": "^2.5.2",
-    "gulp-cssmin": "^0.1.7",
-    "gulp-if": "^2.0.1",
-    "gulp-imagemin": "^2.4.0",
-    "gulp-replace": "^0.5.4",
-    "gulp-rev": "^7.0.0",
+    "browser-sync": "^2.18.13",
+    "browserify": "^14.5.0",
+    "gulp": "^3.9.1",
+    "gulp-cache": "^1.0.1",
+    "gulp-clean": "^0.3.2",
+    "gulp-cssmin": "^0.2.0",
+    "gulp-if": "^2.0.2",
+    "gulp-imagemin": "^4.0.0",
+    "gulp-jsdoc3": "^1.0.1",
+    "gulp-rev": "^8.1.0",
     "gulp-rev-replace": "^0.4.3",
-    "gulp-uglify": "^1.2.0",
-    "gulp-util": "^3.0.7",
-    "imagemin-pngquant": "^4.2.2",
-    "requirejs": "^2.3.5",
-    "run-sequence": "^1.1.0",
-    "through2": "^0.6.5"
+    "gulp-uglify": "^3.0.0",
+    "imagemin-pngquant": "^5.0.1",
+    "node-sass": "^4.7.2",
+    "run-sequence": "^2.2.0",
+    "vinyl-source-stream": "^1.1.0",
+    "zepto-modules": "^1.0.1"
   },
   "dependencies": {
-    "browser-sync": "^2.11.2",
-    "docdash": "^0.4.0",
-    "gulp-jsdoc3": "^1.0.1",
-    "node-sass": "^4.5.3"
+    "babel-core": "^6.26.0",
+    "babel-preset-es2015": "^6.24.1",
+    "gulp-babel": "^7.0.0"
   }
 }
